feat: useUserAvailability hook 수정
