+++ frontend/src/hooks/useUserAvailability.ts
@@ -1,7 +1,12 @@
-import { createUserAvailableTime } from '@/apis/time/time';
+import { getUserAvailableTime, updateUserAvailableTime } from '@/apis/time/time';
 import { toCreateUserAvailability } from '@/apis/transform/toCreateUserAvailablity';
 import { UserAvailability } from '@/types/userAvailability';
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
+
+const initialUserAvailability = {
+  userName: '앙부일구',
+  selectedTimes: new Set<string>(),
+};
 
 export const useUserAvailability = ({
   name,
@@ -10,10 +15,10 @@ export const useUserAvailability = ({
   name: string;
   session: string | null;
 }) => {
-  const initialUserAvailability = {
-    userName: '메이토',
-    selectedTimes: new Set<string>(),
-  };
+  useEffect(() => {
+    fetchUserAvailableTime();
+  }, []);
+
   const [userAvailability, setUserAvailability] =
     useState<UserAvailability>(initialUserAvailability);
 
@@ -31,13 +36,29 @@ export const useUserAvailability = ({
   const userAvailabilitySubmit = async () => {
     try {
       const payload = toCreateUserAvailability(userAvailability);
-      const response = await createUserAvailableTime(session, payload);
-      console.log('User availability submitted successfully:', response);
+      const response = await updateUserAvailableTime(session, payload);
       return response;
     } catch (err) {
       const e = err as Error;
       alert(e.message);
       console.error(err);
+    }
+  };
+
+  const fetchUserAvailableTime = async () => {
+    if (!session) {
+      alert('세션이 없습니다. 다시 시도해주세요.');
+      return;
+    }
+
+    const userAvailableTimeInfo = await getUserAvailableTime(session, name);
+
+    if (userAvailableTimeInfo.timeSlots.length > 0) {
+      const selectedTimesResponse = new Set(
+        userAvailableTimeInfo.timeSlots.map((item) => item.dateTime)
+      );
+      selectedTimes.set(selectedTimesResponse);
+      userName.set(name);
     }
   };
